Expand & refactor Expression interfaces to support more expressions
