--- svalbard/odata/src/main/java/org/n52/svalbard/odata/expr/ArithmeticExpr.java
@@ -1,10 +1,0 @@
-package org.n52.svalbard.odata.expr;
-
-/**
- * Interface to denote that Expression can possibly be used in arithmetic operations.
- * Does not guarantee that the result will be semantically correct!
- *
- * @author <a href="mailto:j.speckamp@52north.org">Jan Speckamp</a>
- */
-public interface ArithmeticExpr extends Expr {
-}
